Interpolate Volume from Time

Interpolates the volume column values, based on the time column values.

chrom_interp_volume(.data, time, volume)

Arguments

.data

Data frame with a volume column to be interpolated.

time

Name of the time column to use for interpolation.

volume

Name of the volumn column to interpolate along the time.

Value

a [tibble][tibble::tibble-package]

Details

Some FPLC systems don't report accurate volume data, but do report accurate timing data. This function takes the time and volume data and interpolates the otherwise constant volume data in accordance with the time points.

Examples

fl <- system.file(
  "extdata",
  "sec.txt",
  package = "chromr"
)

dat <- fl %>%

  chrom_read_quadtech(interp_volume = FALSE)

dat
#> # A tibble: 3,840 × 9
#>     time volume    wl       abs gradient_pump    uv conductivity gp_pr…¹ channel
#>    <dbl>  <dbl> <dbl>     <dbl>         <dbl> <dbl>        <dbl>   <dbl>   <dbl>
#>  1     0      0   280  0                    0     0         16.0     265       1
#>  2     0      0   260  0                    0     0         16.0     265       2
#>  3     0      0   550  0.00550              0     0         16.0     265       3
#>  4     0      0   650  0                    0     0         16.0     265       4
#>  5     1      0   280  0.000004             0     0         16.0     266       1
#>  6     1      0   260  0.000013             0     0         16.0     266       2
#>  7     1      0   550 -0.00016              0     0         16.0     266       3
#>  8     1      0   650  0.00188              0     0         16.0     266       4
#>  9     2      0   280  0                    0     0         16.0     270       1
#> 10     2      0   260 -0.000008             0     0         16.0     270       2
#> # … with 3,830 more rows, and abbreviated variable name ¹​gp_pressure
#> # ℹ Use `print(n = ...)` to see more rows

dat %>%

  chrom_interp_volume(time, volume)
#> # A tibble: 960 × 2
#>     time volume
#>    <dbl>  <dbl>
#>  1     0   0.01
#>  2     1   0.02
#>  3     2   0.03
#>  4     3   0.04
#>  5     4   0.05
#>  6     5   0.06
#>  7     6   0.07
#>  8     7   0.08
#>  9     8   0.09
#> 10     9   0.1 
#> # … with 950 more rows
#> # ℹ Use `print(n = ...)` to see more rows